Update About.tsx
Changes and Explanation:
Random Hover Colors:

I created a getRandomColor() function that generates a random hex color (#RRGGBB format).
This function is used to generate a list of random colors, which are stored in the hoverColors state using the useState hook.
State Initialization:

The useEffect hook initializes the random colors when the component mounts by generating a new set of colors based on the number of milestones. This ensures the colors are set once when the component loads.
Dynamic Application of Hover Colors:

For each milestone, I applied a random color to the background of the paragraph container (div that wraps the text) by passing the corresponding random color from hoverColors.
Hover Effect:

I added a hover effect (group-hover:text-indigo-600), which changes the text color when the user hovers over the event's text. The random color is applied to the background.
Result: