Explanation of Changes: Anime.js Hover Effect:
When the user hovers over the "View My Work" button, it scales up and becomes fully opaque. This is done by using the anime() function and adjusting the scale and opacity properties.
When the hover ends, the button reverts to its original state with reduced opacity and scale.
Looping Animation:

A background animation has been added to the button that constantly scales the button up and down between scale: [1, 1.1] with a fade effect (opacity: [1, 0.7]). This gives the button a subtle but continuous motion to grab attention.
Responsive & Smooth Animations:

The animations are smooth, and the transitions are fluid due to the use of easing properties such as easeInOutQuad, easeOutQuad, and easeInQuad to make the hover and normal state transitions feel more natural.