Straight Line Path:
The path has been modified to a straight line: d="M10,25 L190,25". This creates a horizontal straight line starting from point (10, 25) and ending at (190, 25).
Vivus Animation:

The Vivus animation with the oneByOne type still applies, meaning the SVG path will be drawn one segment at a time, but now it will animate a straight line instead of a curve.
strokeDasharray:

The strokeDasharray is set to 180 to match the length of the straight line path. You can adjust this value if the line length changes.